read Protect

+++ configuration_file/configuration.hpp
@@ -19,7 +19,7 @@ struct server
 {
     int socket_id;
     std::string root, host, port, server_name;
-    unsigned int mx_cl_bd_size;
+    size_t mx_cl_bd_size;
     std::map<std::string, std::string> err_pages;
     std::deque<location> locations;
 };
